Fix all remaining doc- and test-issues
