--- docs/api_docs/solvers/markov.md
@@ -1,1 +1,0 @@
-::: probdiffeq.solvers.markov
